requirements.txt
pandas
openai
python-dotenv
scikit-learn
numpy
requests
flask